Make fixed smsUrl
