+++ eisbuk-admin/functions/src/constants.ts
@@ -1,0 +1,3 @@
+export const __functionsZone__ = "europe-west6";
+
+export const __smsUrl__ = "https://gatewayapi.com/rest/mtsms";
