+++ eisbuk-shared/src/enums/errorMessages.ts
@@ -7,7 +7,6 @@ export enum HTTPSErrors {
 }
 
 export enum SendSMSErrors {
-  NoProviderURL = "No URL for SMS sending service provider found, add your SMS service provider URL to 'smsConfig.url' for organization",
   NoAuthToken = "No auth token for SMS sending service provider found, add your SMS service provider auth token 'smsConfig.authToken' for organization",
   SendingFailed = "SMS sending failed on provider's side. Check the details.",
 }
